requirements.txt
h5py==3.9.0
imageio==2.31.6
kaggle==1.5.16
kagglehub==0.2.2
keras==2.15.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.10
MedPy==0.4.0
mdit-py-plugins==0.4.0
numpy==1.23.5
pandas==2.0.3
Pillow==9.4.0
SimpleITK==2.3.1
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.4
sklearn-pandas==2.2.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-gcs-config==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.23.0
tqdm==4.65.0
streamlit












